+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Changed
+- Adapter class now handles reading and modifying domain records.
+- Adapter class has been split into resource and relationship adapters.
+
+### Removed
+- Hydrators have been removed as a result of the adapter changes described above.
+
 ## [0.11.0] - 2017-09-02
 
 ### Removed